update: split demo.py into a WebSearcher/demos package
