+++ WebSearcher/demos/__main__.py
@@ -1,0 +1,6 @@
+"""Enable `python -m WebSearcher.demos` to dispatch the demo CLI."""
+
+from .cli import main
+
+if __name__ == "__main__":
+    main()
