+++ pyproject.toml
@@ -24,7 +24,7 @@ homepage = "http://github.com/gitronald/WebSearcher"
 repository = "http://github.com/gitronald/WebSearcher"
 
 [project.scripts]
-ws-demo = 'WebSearcher.demo:main'
+ws-demo = 'WebSearcher.demos:main'
 
 [dependency-groups]
 dev = [
@@ -91,6 +91,6 @@ build-backend = "hatchling.build"
 # repo but is kept out of the distributed source archive. pyproject.toml,
 # README.md, and LICENSE are added automatically as project metadata.
 only-include = [
-    "WebSearcher",  # the importable package (includes the demos at WebSearcher/demo.py)
+    "WebSearcher",  # the importable package (includes the demos at WebSearcher/demos/)
     "CHANGELOG.md",
 ]
